+++ ExtendedTree/tree.cpp
@@ -69,7 +69,7 @@ void precompute_dir_layout(const std::string &dir, FileNode &parent, Stats &stat
     }
 }
 
-void traverse_dirs_print_relative(const std::unique_ptr<FileNode> &node, int total_size, int depth = 0)
+void traverse_dirs_print_relative(const std::unique_ptr<FileNode> &node, uintmax_t total_size, int depth = 0)
 {
     node->print(depth, total_size);
     depth++;
@@ -89,7 +89,7 @@ void traverse_dirs_print_absolute(const std::unique_ptr<FileNode> &node, int dep
     }
 }
 
-void traverse_dirs_print_relative_dirs(const std::unique_ptr<FileNode> &node, int total_size, int depth = 0)
+void traverse_dirs_print_relative_dirs(const std::unique_ptr<FileNode> &node, uintmax_t total_size, int depth = 0)
 {
     node->print_dirs_only(depth, total_size);
     depth++;
